Add files via upload
tweak to the way that asteroid belts are handled.  belts are now numbered sequentially, rather than with the number of the planet they take the place of.  working on a way to move the belt details into the planet block itself, but that means moving all the echo commands to adding text to an array.  hindsight: shoulda done it this way from the beginning, but rusty programmer and all.